Move some file from checker to checks namespace.
