+++ osu.Game.Rulesets.Karaoke/Edit/Checks/LyricCheckerConfig.cs
@@ -4,7 +4,7 @@
 using osu.Game.Rulesets.Karaoke.Edit.Checker.Types;
 using osu.Game.Rulesets.Karaoke.Utils;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Checker.Lyrics
+namespace osu.Game.Rulesets.Karaoke.Edit.Checks
 {
     public class LyricCheckerConfig : IHasConfig<LyricCheckerConfig>
     {
